+++ tools/tests/data/fake_stac.json
