+++ tools/tests/test_s2_gap_report.py
@@ -3,9 +3,9 @@ from unittest.mock import patch
 import boto3
 from moto import mock_s3
 
-from monitoring.tests.conftest import *
-from monitoring.tools import s2_gap_report
-from monitoring.tools.s2_gap_report import (
+from tools.tests.conftest import *
+from tools.monitoring import s2_gap_report
+from tools.monitoring.s2_gap_report import (
     get_and_filter_cogs_keys,
     generate_buckets_diff,
 )
