+++ tools/tests/test_tools.py
@@ -4,9 +4,9 @@ from moto import mock_s3, mock_sqs
 from odc.aws.queue import publish_message
 from urlpath import URL
 
-from monitoring.tests.conftest import REGION, TEST_BUCKET_NAME
-from monitoring.tools.check_dead_queues import check_deadletter_queues
-from monitoring.tools.utils import (
+from tools.tests.conftest import REGION, TEST_BUCKET_NAME
+from tools.monitoring.check_dead_queues import check_deadletter_queues
+from tools.monitoring.utils import (
     find_latest_report,
     read_report,
     split_list_equally,
